[UHYU-90] chore: 시크릿 파일 실수로 추가로 인한 커밋버전 되돌리기(develop 머지 직후)

+++ src/main/java/com/ureca/uhyu/domain/user/entity/User.java
@@ -29,7 +29,7 @@ public class User extends BaseEntity {
     @Column(length = 200, unique = true)
     private String email;
 
-    private Byte age;
+    private Integer age;
 
     @Enumerated(EnumType.STRING)
     private Gender gender;
